Organize unit tests under src/tests/unit folder
- Create src/tests/unit/ directory structure
- Move src/App.test.tsx to src/tests/unit/App.test.tsx
- Move src/pages/AccountsPage.test.tsx to src/tests/unit/AccountsPage.test.tsx
- Update import paths in unit tests to reflect new structure
- Remove old unit test files from their previous locations

This creates a cleaner test organization:
- src/tests/unit/ - Unit tests
- src/tests/integration/ - Integration tests

npm scripts continue to work unchanged due to pattern-based matching.
